refactor:  컴포넌트 - 컴포넌트명 변경 MainFooter -> Footer - 불필요한 태그 제거 - css 색상 상수화
