+++ frontend/src/App.css
@@ -1,8 +1,26 @@
+:root {
+  --gray-002: #eee;
+  --gray-003: #ddd;
+  --gray-004: #ccc;
+  --gray-005: #bbb;
+  --gray-006: #aaa;
+  --gray-007: #999;
+  --gray-008: #666;
+  --black: #333;
+  --white: #fff;
+  --light-blue-gray: #f1f2f6;
+  --blue: #279cf6;
+  --pc-main-width: 800px;
+  --pc-main-width-narrow: 512px;
+}
+
 * {
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI",
     "Apple SD Gothic Neo", Roboto, "Helvetica Neue", Arial, sans-serif,
     "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 
 html,
@@ -10,27 +28,20 @@ body {
   margin: 0;
   padding: 0;
   height: 100%;
-  background: #ced6e0;
+  background: var(--light-blue-gray);
+}
+
+a {
+  text-decoration: none;
 }
 
 #root {
+  height: 100%;
   display: flex;
   flex-direction: column;
-  height: 100%;
 }
 
-.main-view {
-  flex: 1 0 auto;
-}
-
-button {
-  background-color: #ffffff;
-  border-radius: 10px;
-  font-size: 16px;
-  border: 0;
-  outline: 0;
-}
-
-button:active {
-  background-color: #cccccc;
+.main {
+  flex: 1;
+  padding: 2.5rem 0;
 }
