+++ frontend/src/components/Footer/Footer.stories.js
@@ -1,0 +1,13 @@
+import React from "react";
+import Footer from "./Footer";
+
+export default {
+  title: "Footer",
+  component: Footer,
+};
+
+const Template = (args) => <Footer {...args} />;
+
+export const Default = Template.bind({});
+
+Default.args = {};
