--- frontend/src/components/MainFooter/MainFooter.module.css
@@ -1,62 +1,0 @@
-.main-footer {
-  user-select: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  padding: 30px 0;
-  background: #333;
-  color: #fff;
-  flex: 0 0 auto;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 900px;
-}
-
-.logo-title {
-  display: flex;
-  align-items: center;
-}
-
-.logo {
-  width: 200px;
-}
-
-.title {
-  font-size: 22px;
-}
-
-a {
-  text-decoration: none;
-  color: #fff;
-}
-
-.social {
-  display: flex;
-  margin: 20px 0 0 10px;
-}
-
-.social a {
-  display: flex;
-  width: 46px;
-  height: 46px;
-  justify-content: center;
-  align-items: center;
-  border-radius: 100%;
-  border: 1px solid #fff;
-  transition: all 0.2s linear;
-  margin-right: 10px;
-}
-
-.social a:hover {
-  color: #333;
-  background-color: #fff;
-}
-
-.copyright {
-  margin: 10px;
-}
