--- frontend/src/components/MainFooter/MainFooter.stories.js
@@ -1,13 +1,0 @@
-import React from "react";
-import MainFooter from "./MainFooter";
-
-export default {
-  title: "MainFooter",
-  component: MainFooter,
-};
-
-const Template = (args) => <MainFooter {...args} />;
-
-export const Default = Template.bind({});
-
-Default.args = {};
